Add SCRIPT_NAME to README
Thanks to switching to gunicorn, SCRIPT_NAME env var is now honoured.
This allows to run the app under a prefixed URL which in turns allows for
easier reverse-proxying.
Note that the prefix must not be removed from the URL by the reverse proxy

Example conf for nginx:
location /gptwol/ {
        proxy_pass http://127.0.0.1:8080/gptwol/;
        proxy_redirect off;
}

Signed-off-by: Nicolas Morey <nicolas@morey.ovh>